pyproject.toml
[project]
name = "fitlog"
version = "0.1.0"
description = "FitLog - Fitness and Nutrition Tracking API"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "pydantic>=2.0",
    "uvicorn[standard]>=0.30.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "streamlit>=1.28.0",
    "sqlmodel>=0.0.18",
    "alembic>=1.13.0",
    "aiosqlite>=0.19.0",
    "redis>=5.0.0",
    "celery>=5.4.0",
    "pydantic-ai>=0.1.0",
    "passlib>=1.7.4",
    "python-jose>=3.3.0",
    "bcrypt>=4.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
